fixed searchInCuboid once and for all

+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/world/ModuleChestAura.kt
@@ -169,7 +169,7 @@ object ModuleChestAura : Module("ChestAura", Category.WORLD) {
         val radiusSquared = radius * radius
         val eyesPos = mc.player!!.eyes
 
-        val blocksToProcess = searchBlocksInCuboid(radius.toInt(), eyesPos) { pos, state ->
+        val blocksToProcess = searchBlocksInCuboid(radius, eyesPos) { pos, state ->
             targetedBlocks.contains(state.block) && pos !in clickedBlocks && getNearestPoint(
                 eyesPos,
                 Box(pos, pos.add(1, 1, 1))
