+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/world/ModuleFucker.kt
@@ -146,7 +146,7 @@ object ModuleFucker : Module("Fucker", Category.WORLD) {
         val radiusSquared = radius * radius
         val eyesPos = player.eyes
 
-        val blockToProcess = searchBlocksInCuboid(radius.toInt(), eyesPos) { pos, state ->
+        val blockToProcess = searchBlocksInCuboid(radius, eyesPos) { pos, state ->
             targetedBlocks.contains(state.block) && getNearestPoint(
                 eyesPos, Box(pos, pos.add(1, 1, 1))
             ).squaredDistanceTo(eyesPos) <= radiusSquared
